# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "(Re)Sources Relationnelles",
        "version": "0.26.2"
    },
    "servers": [
        {
            "url": "http://localhost:80/api"
        }
    ],
    "paths": {
...
        "/stats/connections": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Get connection statistics",
                "description": "Fetches connection counts for a given date range along with the day having the highest average connections.",
                "operationId": "getConnections",
                "parameters": [
                    {
                        "name": "startDate",
                        "in": "query",
                        "description": "Start date for fetching connections (format: DD-MM-YYYY)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "endDate",
                        "in": "query",
                        "description": "End date for fetching connections (format: DD-MM-YYYY)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "connections": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "date": {
                                                        "type": "string",
                                                        "example": "06-03-2024"
                                                    },
                                                    "numberConnections": {
                                                        "type": "integer",
                                                        "example": 2
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "average": {
                                            "properties": {
                                                "day": {
                                                    "type": "string",
                                                    "example": "mercredi"
                                                },
                                                "value": {
                                                    "type": "string",
                                                    "example": "2,0"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "BearerAuth": {
                "type": "apiKey",
                "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}